pyproject.toml
[tool.poetry]
name = "thcode"
version = "0.1.0"
description = ""
authors = ["Erik Giorgis <git@egiorg.is>"]
readme = "README.org"

[tool.poetry.dependencies]
python = "^3.10"
git2net = "^1.7.0"
networkx = "^3.1"
jupyter = "^1.0.0"
matplotlib = "^3.7.1"
tabulate = "^0.9.0"
pathpy = "^2.2.0"
pygithub = "^1.59.0"
unidecode = "^1.3.6"
gambit-disambig = "^1.0.3"
seaborn = "^0.12.2"
plotly = "^5.16.1"
statsmodels = "^0.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
